# backend/docs/swagger.yaml
definitions:
  database.Aircraft:
    properties:
      aircraft_id:
        type: integer
      aircraft_manufacturer:
        type: string
      aircraft_name:
        type: string
      aircraft_type:
        type: string
      capacity:
        type: integer
    type: object
  database.Airport:
    properties:
      airport_id:
        type: integer
      city:
        type: string
      country:
        type: string
      iata:
        type: string
      icao:
        type: string
      name:
        type: string
      timezone:
        description: |-
          Latitude  string
          Longitude string
          Altitude  string
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/v1/aircrafts:
    get:
      consumes:
      - application/json
      description: Get all aircrafts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Aircraft'
            type: array
      summary: Get all aircrafts
      tags:
      - aircrafts
    post:
      consumes:
      - application/json
      description: Create an aircraft
      parameters:
      - description: Aircraft object
        in: body
        name: aircraft
        required: true
        schema:
          $ref: '#/definitions/database.Aircraft'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Aircraft'
      summary: Create an aircraft
      tags:
      - aircrafts
  /api/v1/aircrafts/{aircraft_id}:
    delete:
      consumes:
      - application/json
      description: Remove aircraft by ID
      parameters:
      - description: Aircraft ID
        in: path
        name: aircraft_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove aircraft by ID
      tags:
      - aircrafts
    get:
      consumes:
      - application/json
      description: Get aircraft by ID
      parameters:
      - description: Aircraft ID
        in: path
        name: aircraft_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Aircraft'
      summary: Get aircraft by ID
      tags:
      - aircrafts
    put:
      consumes:
      - application/json
      description: Update aircraft by ID
      parameters:
      - description: Aircraft ID
        in: path
        name: aircraft_id
        required: true
        type: integer
      - description: Aircraft object
        in: body
        name: aircraft
        required: true
        schema:
          $ref: '#/definitions/database.Aircraft'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Aircraft'
      summary: Update aircraft by ID
      tags:
      - aircrafts
  /api/v1/aircrafts/search/aircraft/{aircraft_type}:
    get:
      consumes:
      - application/json
      description: Get aircrafts by aircraft type
      parameters:
      - description: Aircraft Type
        in: path
        name: aircraft_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Aircraft'
            type: array
      summary: Get aircrafts by aircraft type
      tags:
      - aircrafts
  /api/v1/aircrafts/search/capacity/{capacity}:
    get:
      consumes:
      - application/json
      description: Get all aircrafts by capacity threshold (>= capacity)
      parameters:
      - description: Aircraft Capacity
        in: path
        name: capacity
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Aircraft'
            type: array
      summary: Get all aircrafts by capacity threshold (>= capacity)
      tags:
      - aircrafts
  /api/v1/aircrafts/search/manufacturer/{aircraft_manufacturer}:
    get:
      consumes:
      - application/json
      description: Get all aircrafts by aircraft manufacturer
      parameters:
      - description: Aircraft Manufacturer
        in: path
        name: aircraft_manufacturer
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Aircraft'
            type: array
      summary: Get all aircrafts by aircraft manufacturer
      tags:
      - aircrafts
  /api/v1/aircrafts/search/name/{aircraft_name}:
    get:
      consumes:
      - application/json
      description: Get all aircrafts by aircraft name
      parameters:
      - description: Aircraft Name
        in: path
        name: aircraft_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Aircraft'
            type: array
      summary: Get all aircrafts by aircraft name
      tags:
      - aircrafts
...
  /api/v1/airports/city/{city}:
    get:
      consumes:
      - application/json
      description: Get airport by city
      parameters:
      - description: City
        in: path
        name: city
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Airport'
            type: array
      summary: Get airport by city
      tags:
      - airports
  /api/v1/airports/iata/{iata}:
    get:
      consumes:
      - application/json
      description: Get airport by IATA code
      parameters:
      - description: IATA code
        in: path
        name: iata
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Airport'
      summary: Get airport by IATA code
      tags:
      - airports
  /api/v1/airports/icao/{icao}:
    get:
      consumes:
      - application/json
      description: Get airport by ICAO code
      parameters:
      - description: ICAO code
        in: path
        name: icao
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Airport'
      summary: Get airport by ICAO code
      tags:
      - airports
  /api/v1/airports/id/{airport_id}:
    get:
      consumes:
      - application/json
      description: Get airport by ID
      parameters:
      - description: Airport ID
        in: path
        name: airport_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Airport'
      summary: Get airport by ID
      tags:
      - airports
...
swagger: "2.0"